uncompress -> decompress, 增加终端输出， 耗时显，

+++ Data_Compressor-linux/huffman.h
@@ -28,7 +28,7 @@ void node_insert(node ** head, node* ins);
 
 node* pop(node** head);
 
-node* generate_huffman_tree(int* char_table, int tnum);		// 生成huffman树，返回树root结点
+node* generate_huffman_tree(int* char_table, long long tnum);		// 生成huffman树，返回树root结点
 double generate_symbole_table(node* head, int* ar, data_buffer* symbols);			// 生成符号表
 
 void output_extra_bit_len(data_buffer* symbols, int* table);	// 计算最后一个缓冲区的位数，并输出
